Refactor perform FVS versioning work in dedicated class
